delete rg

+++ steps/azure/azureDeleteResourceGroup/onExecute.sh
@@ -1,10 +1,14 @@
-create_resource_group() {
+delete_resource_group() {
   echo "onExecute step"
     printenv
 
     # Create resource group
-    az group delete --name $res_aksClusterResource_azureResourceGroup --no-wait --yes
-
+    if [ $res_aksClusterResource_isDeleted = true ]
+      then
+        az group delete --name $res_aksClusterResource_azureResourceGroup --no-wait --yes
+      else
+        echo "The cluster is not deleted yet"
+    fi
 }
 
-execute_command create_resource_group
+execute_command delete_resource_group
